ДЗ 5.12. Модуляция
